chore: split up tests classes to reduce length
