+++ test/BuildingRegistry.Tests/ProjectionTests/Integration/Building/BuildingVersionProjectionsTests-LegacyEvents.cs
@@ -1,0 +1,6 @@
+﻿namespace BuildingRegistry.Tests.ProjectionTests.Integration.Building
+{
+    public partial class BuildingVersionProjectionsTests
+    {
+    }
+}
