+++ test/BuildingRegistry.Tests/ProjectionTests/Integration/BuildingUnit/BuildingUnitVersionProjectionsTests-LegacyBuildingEvents.cs
@@ -1,0 +1,7 @@
+﻿namespace BuildingRegistry.Tests.ProjectionTests.Integration.BuildingUnit
+{
+    public partial class BuildingUnitVersionProjectionsTests
+    {
+
+    }
+}
