+++ test/BuildingRegistry.Tests/ProjectionTests/Integration/BuildingUnit/BuildingUnitVersionProjectionsTests.cs
@@ -1,5 +1,6 @@
-namespace BuildingRegistry.Tests.ProjectionTests.Integration
+﻿namespace BuildingRegistry.Tests.ProjectionTests.Integration.BuildingUnit
 {
+    using System;
     using System.Collections.Generic;
     using System.Linq;
     using System.Threading.Tasks;
@@ -8,29 +9,35 @@ namespace BuildingRegistry.Tests.ProjectionTests.Integration
     using Be.Vlaanderen.Basisregisters.GrAr.Provenance;
     using Be.Vlaanderen.Basisregisters.ProjectionHandling.SqlStreamStore;
     using Be.Vlaanderen.Basisregisters.Utilities.HexByteConvertor;
-    using Building;
-    using Building.Events;
-    using Extensions;
-    using Fixtures;
+    using BuildingRegistry.Building;
+    using BuildingRegistry.Building.Events;
+    using BuildingRegistry.Projections.Integration;
+    using BuildingRegistry.Projections.Integration.BuildingUnit.Version;
+    using BuildingRegistry.Projections.Integration.Converters;
+    using BuildingRegistry.Projections.Integration.Infrastructure;
+    using BuildingRegistry.Tests.Extensions;
+    using BuildingRegistry.Tests.Fixtures;
+    using BuildingRegistry.Tests.Legacy.Autofixture;
     using FluentAssertions;
     using Microsoft.Extensions.Options;
+    using Moq;
     using NetTopologySuite.IO;
-    using Projections.Integration.BuildingUnit.LatestItem;
-    using Projections.Integration.Converters;
-    using Projections.Integration.Infrastructure;
-    using Tests.Legacy.Autofixture;
     using Xunit;
+    using IAddresses = Projections.Integration.IAddresses;
 
-    public class BuildingUnitLatestItemProjectionsTests : IntegrationProjectionTest<BuildingUnitLatestItemProjections>
+    public partial class BuildingUnitVersionProjectionsTests : IntegrationProjectionTest<BuildingUnitVersionProjections>
     {
         private const string BuildingNamespace = "https://data.vlaanderen.be/id/gebouw";
         private const string BuildingUnitNamespace = "https://data.vlaanderen.be/id/gebouweenheid";
 
         private readonly Fixture _fixture;
         private readonly WKBReader _wkbReader = WKBReaderFactory.Create();
+        private readonly Mock<IPersistentLocalIdFinder> _persistentLocalIdFinder;
 
-        public BuildingUnitLatestItemProjectionsTests()
+        public BuildingUnitVersionProjectionsTests()
         {
+            _persistentLocalIdFinder = new Mock<IPersistentLocalIdFinder>();
+
             _fixture = new Fixture();
             _fixture.Customizations.Add(new WithUniqueInteger());
             _fixture.Customize(new InfrastructureCustomization());
@@ -65,37 +72,34 @@ namespace BuildingRegistry.Tests.ProjectionTests.Integration
                 .Given(new Envelope<BuildingWasMigrated>(new Envelope(buildingWasMigrated, metadata)))
                 .Then(context =>
                 {
-                    var buildingUnits = context.BuildingUnitLatestItems
+                    var buildingUnits = context.BuildingUnitVersions
                         .Where(x => x.BuildingPersistentLocalId == buildingWasMigrated.BuildingPersistentLocalId)
                         .ToList();
 
                     foreach (var buildingUnit in buildingWasMigrated.BuildingUnits)
                     {
-                        var buildingUnitLatestItem = buildingUnits
+                        var buildingUnitVersion = buildingUnits
                             .Single(x => x.BuildingUnitPersistentLocalId == buildingUnit.BuildingUnitPersistentLocalId);
 
-                        buildingUnitLatestItem.BuildingPersistentLocalId.Should().Be(buildingWasMigrated.BuildingPersistentLocalId);
-                        buildingUnitLatestItem.OsloStatus.Should().Be(BuildingUnitStatus.Parse(buildingUnit.Status).Map());
-                        buildingUnitLatestItem.Status.Should().Be(buildingUnit.Status);
-                        buildingUnitLatestItem.OsloFunction.Should().Be(BuildingUnitFunction.Parse(buildingUnit.Function).Map());
-                        buildingUnitLatestItem.Function.Should().Be(buildingUnit.Function);
-                        buildingUnitLatestItem.OsloGeometryMethod.Should().Be(BuildingUnitPositionGeometryMethod.Parse(buildingUnit.GeometryMethod).Map());
-                        buildingUnitLatestItem.GeometryMethod.Should().Be(buildingUnit.GeometryMethod);
-                        buildingUnitLatestItem.Geometry.Should().BeEquivalentTo(_wkbReader.Read(buildingUnit.ExtendedWkbGeometry.ToByteArray()));
-                        buildingUnitLatestItem.HasDeviation.Should().BeFalse();
-                        buildingUnitLatestItem.IsRemoved.Should().Be(buildingUnit.IsRemoved);
-                        buildingUnitLatestItem.Namespace.Should().Be(BuildingUnitNamespace);
-                        buildingUnitLatestItem.Puri.Should().Be($"{BuildingUnitNamespace}/{buildingUnitLatestItem.BuildingUnitPersistentLocalId}");
-                        buildingUnitLatestItem.VersionTimestamp.Should().Be(buildingWasMigrated.Provenance.Timestamp);
+                        buildingUnitVersion.BuildingPersistentLocalId.Should().Be(buildingWasMigrated.BuildingPersistentLocalId);
+                        buildingUnitVersion.Status.Should().Be(BuildingUnitStatus.Parse(buildingUnit.Status).Status);
+                        buildingUnitVersion.OsloStatus.Should().Be(BuildingUnitStatus.Parse(buildingUnit.Status).Map());
+                        buildingUnitVersion.Function.Should().Be(BuildingUnitFunction.Parse(buildingUnit.Function).Function);
+                        buildingUnitVersion.OsloFunction.Should().Be(BuildingUnitFunction.Parse(buildingUnit.Function).Map());
+                        buildingUnitVersion.GeometryMethod.Should().Be(BuildingUnitPositionGeometryMethod.Parse(buildingUnit.GeometryMethod).GeometryMethod);
+                        buildingUnitVersion.OsloGeometryMethod.Should().Be(BuildingUnitPositionGeometryMethod.Parse(buildingUnit.GeometryMethod).Map());
+                        buildingUnitVersion.Geometry.Should().BeEquivalentTo(_wkbReader.Read(buildingUnit.ExtendedWkbGeometry.ToByteArray()));
+                        buildingUnitVersion.HasDeviation.Should().BeFalse();
+                        buildingUnitVersion.IsRemoved.Should().Be(buildingUnit.IsRemoved);
+                        buildingUnitVersion.Namespace.Should().Be(BuildingUnitNamespace);
+                        buildingUnitVersion.PuriId.Should().Be($"{BuildingUnitNamespace}/{buildingUnitVersion.BuildingUnitPersistentLocalId}");
+                        buildingUnitVersion.VersionTimestamp.Should().Be(buildingWasMigrated.Provenance.Timestamp);
+                        buildingUnitVersion.CreatedOnTimestamp.Should().Be(buildingWasMigrated.Provenance.Timestamp);
 
-                        var buildingUnitAddresses = context.BuildingUnitAddresses
-                            .Where(x => x.BuildingUnitPersistentLocalId == buildingUnitLatestItem.BuildingUnitPersistentLocalId)
-                            .ToList();
-
-                        buildingUnitAddresses.Should().HaveCount(buildingUnit.AddressPersistentLocalIds.Count);
+                        buildingUnitVersion.Addresses.Should().HaveCount(3);
                         foreach (var addressPersistentLocalId in buildingUnit.AddressPersistentLocalIds)
                         {
-                            buildingUnitAddresses.SingleOrDefault(x => x.AddressPersistentLocalId == addressPersistentLocalId)
+                            buildingUnitVersion.Addresses.SingleOrDefault(x => x.AddressPersistentLocalId == addressPersistentLocalId)
                                 .Should().NotBeNull();
                         }
                     }
@@ -111,7 +115,12 @@ namespace BuildingRegistry.Tests.ProjectionTests.Integration
             _fixture.Customize(new WithFixedBuildingUnitPersistentLocalId());
 
             var buildingUnitWasPlannedV2 = _fixture.Create<BuildingUnitWasPlannedV2>();
-            var buildingOutlineWasChanged = _fixture.Create<BuildingOutlineWasChanged>();
+            var buildingOutlineWasChanged = new BuildingOutlineWasChanged(
+                _fixture.Create<BuildingPersistentLocalId>(),
+                new []{ _fixture.Create<BuildingUnitPersistentLocalId>() },
+                _fixture.Create<ExtendedWkbGeometry>(),
+                _fixture.Create<ExtendedWkbGeometry>());
+            ((ISetProvenance)buildingOutlineWasChanged).SetProvenance(_fixture.Create<Provenance>());
 
             var position = _fixture.Create<long>();
 
@@ -132,15 +141,15 @@ namespace BuildingRegistry.Tests.ProjectionTests.Integration
                     new Envelope<BuildingOutlineWasChanged>(new Envelope(buildingOutlineWasChanged, buildingOutLineWasChangedMetadata)))
                 .Then(async context =>
                 {
-                    var buildingUnitLatestItem =
-                        await context.BuildingUnitLatestItems.FindAsync(buildingUnitWasPlannedV2.BuildingUnitPersistentLocalId);
-                    buildingUnitLatestItem.Should().NotBeNull();
+                    var buildingUnitVersion =
+                        await context.BuildingUnitVersions.FindAsync(position + 1, buildingUnitWasPlannedV2.BuildingUnitPersistentLocalId);
+                    buildingUnitVersion.Should().NotBeNull();
 
-                    buildingUnitLatestItem!.OsloGeometryMethod.Should().Be(BuildingUnitPositionGeometryMethod.DerivedFromObject.Map());
-                    buildingUnitLatestItem.GeometryMethod.Should().Be("DerivedFromObject");
-                    buildingUnitLatestItem.Geometry.Should().BeEquivalentTo(
+                    buildingUnitVersion!.GeometryMethod.Should().Be(BuildingUnitPositionGeometryMethod.DerivedFromObject.GeometryMethod);
+                    buildingUnitVersion.OsloGeometryMethod.Should().Be(BuildingUnitPositionGeometryMethod.DerivedFromObject.Map());
+                    buildingUnitVersion.Geometry.Should().BeEquivalentTo(
                         _wkbReader.Read(buildingOutlineWasChanged.ExtendedWkbGeometryBuildingUnits!.ToByteArray()));
-                    buildingUnitLatestItem.VersionTimestamp.Should().Be(buildingOutlineWasChanged.Provenance.Timestamp);
+                    buildingUnitVersion.VersionTimestamp.Should().Be(buildingOutlineWasChanged.Provenance.Timestamp);
                 });
         }
 
@@ -148,38 +157,61 @@ namespace BuildingRegistry.Tests.ProjectionTests.Integration
         public async Task WhenBuildingWasMeasured()
         {
             _fixture.Customize(new WithFixedBuildingPersistentLocalId());
-            _fixture.Customize(new WithFixedBuildingUnitPersistentLocalId());
 
-            var buildingUnitWasPlannedV2 = _fixture.Create<BuildingUnitWasPlannedV2>();
-            var buildingWasMeasured = _fixture.Create<BuildingWasMeasured>();
+            var firstBuildingUnitWasPlannedV2 = _fixture.Create<BuildingUnitWasPlannedV2>();
+            var secondBuildingUnitWasPlannedV2 = _fixture.Create<BuildingUnitWasPlannedV2>();
+            var buildingWasMeasured = new BuildingWasMeasured(
+                _fixture.Create<BuildingPersistentLocalId>(),
+                new []{ new BuildingUnitPersistentLocalId(firstBuildingUnitWasPlannedV2.BuildingUnitPersistentLocalId) },
+                new []{ new BuildingUnitPersistentLocalId(secondBuildingUnitWasPlannedV2.BuildingUnitPersistentLocalId) },
+                _fixture.Create<ExtendedWkbGeometry>(),
+                _fixture.Create<ExtendedWkbGeometry>());
+            ((ISetProvenance)buildingWasMeasured).SetProvenance(_fixture.Create<Provenance>());
 
             var position = _fixture.Create<long>();
 
-            var buildingUnitWasPlannedMetadata = new Dictionary<string, object>
+            var firstBuildingUnitWasPlannedMetadata = new Dictionary<string, object>
             {
-                { AddEventHashPipe.HashMetadataKey, buildingUnitWasPlannedV2.GetHash() },
+                { AddEventHashPipe.HashMetadataKey, firstBuildingUnitWasPlannedV2.GetHash() },
                 { Envelope.PositionMetadataKey, position }
+            };
+            var secondBuildingUnitWasPlannedMetadata = new Dictionary<string, object>
+            {
+                { AddEventHashPipe.HashMetadataKey, firstBuildingUnitWasPlannedV2.GetHash() },
+                { Envelope.PositionMetadataKey, position + 1 }
             };
             var buildingWasMeasuredMetadata = new Dictionary<string, object>
             {
                 { AddEventHashPipe.HashMetadataKey, buildingWasMeasured.GetHash() },
-                { Envelope.PositionMetadataKey, position + 1 }
+                { Envelope.PositionMetadataKey, position + 2 }
             };
+
             await Sut
                 .Given(
-                    new Envelope<BuildingUnitWasPlannedV2>(new Envelope(buildingUnitWasPlannedV2, buildingUnitWasPlannedMetadata)),
+                    new Envelope<BuildingUnitWasPlannedV2>(new Envelope(firstBuildingUnitWasPlannedV2, firstBuildingUnitWasPlannedMetadata)),
+                    new Envelope<BuildingUnitWasPlannedV2>(new Envelope(secondBuildingUnitWasPlannedV2, secondBuildingUnitWasPlannedMetadata)),
                     new Envelope<BuildingWasMeasured>(new Envelope(buildingWasMeasured, buildingWasMeasuredMetadata)))
                 .Then(async context =>
                 {
-                    var buildingUnitLatestItem =
-                        await context.BuildingUnitLatestItems.FindAsync(buildingUnitWasPlannedV2.BuildingUnitPersistentLocalId);
-                    buildingUnitLatestItem.Should().NotBeNull();
+                    var firstBuildingUnitVersion =
+                        await context.BuildingUnitVersions.FindAsync(position + 2, firstBuildingUnitWasPlannedV2.BuildingUnitPersistentLocalId);
+                    firstBuildingUnitVersion.Should().NotBeNull();
 
-                    buildingUnitLatestItem!.OsloGeometryMethod.Should().Be(BuildingUnitPositionGeometryMethod.DerivedFromObject.Map());
-                    buildingUnitLatestItem.GeometryMethod.Should().Be("DerivedFromObject");
-                    buildingUnitLatestItem.Geometry.Should().BeEquivalentTo(
+                    firstBuildingUnitVersion!.GeometryMethod.Should().Be(BuildingUnitPositionGeometryMethod.DerivedFromObject.GeometryMethod);
+                    firstBuildingUnitVersion.OsloGeometryMethod.Should().Be(BuildingUnitPositionGeometryMethod.DerivedFromObject.Map());
+                    firstBuildingUnitVersion.Geometry.Should().BeEquivalentTo(
                         _wkbReader.Read(buildingWasMeasured.ExtendedWkbGeometryBuildingUnits!.ToByteArray()));
-                    buildingUnitLatestItem.VersionTimestamp.Should().Be(buildingWasMeasured.Provenance.Timestamp);
+                    firstBuildingUnitVersion.VersionTimestamp.Should().Be(buildingWasMeasured.Provenance.Timestamp);
+
+                    var secondBuildingUnitVersion =
+                        await context.BuildingUnitVersions.FindAsync(position + 2, secondBuildingUnitWasPlannedV2.BuildingUnitPersistentLocalId);
+                    secondBuildingUnitVersion.Should().NotBeNull();
+
+                    secondBuildingUnitVersion!.GeometryMethod.Should().Be(BuildingUnitPositionGeometryMethod.DerivedFromObject.GeometryMethod);
+                    secondBuildingUnitVersion.OsloGeometryMethod.Should().Be(BuildingUnitPositionGeometryMethod.DerivedFromObject.Map());
+                    secondBuildingUnitVersion.Geometry.Should().BeEquivalentTo(
+                        _wkbReader.Read(buildingWasMeasured.ExtendedWkbGeometryBuildingUnits!.ToByteArray()));
+                    secondBuildingUnitVersion.VersionTimestamp.Should().Be(buildingWasMeasured.Provenance.Timestamp);
                 });
         }
 
@@ -187,39 +219,61 @@ namespace BuildingRegistry.Tests.ProjectionTests.Integration
         public async Task WhenBuildingMeasurementWasCorrected()
         {
             _fixture.Customize(new WithFixedBuildingPersistentLocalId());
-            _fixture.Customize(new WithFixedBuildingUnitPersistentLocalId());
 
-            var buildingUnitWasPlannedV2 = _fixture.Create<BuildingUnitWasPlannedV2>();
-            var buildingMeasurementWasCorrected = _fixture.Create<BuildingMeasurementWasCorrected>();
+            var firstBuildingUnitWasPlannedV2 = _fixture.Create<BuildingUnitWasPlannedV2>();
+            var secondBuildingUnitWasPlannedV2 = _fixture.Create<BuildingUnitWasPlannedV2>();
+            var buildingMeasurementWasCorrected = new BuildingMeasurementWasCorrected(
+                _fixture.Create<BuildingPersistentLocalId>(),
+                new []{ new BuildingUnitPersistentLocalId(firstBuildingUnitWasPlannedV2.BuildingUnitPersistentLocalId) },
+                new []{ new BuildingUnitPersistentLocalId(secondBuildingUnitWasPlannedV2.BuildingUnitPersistentLocalId) },
+                _fixture.Create<ExtendedWkbGeometry>(),
+                _fixture.Create<ExtendedWkbGeometry>());
+            ((ISetProvenance)buildingMeasurementWasCorrected).SetProvenance(_fixture.Create<Provenance>());
 
             var position = _fixture.Create<long>();
 
-            var buildingUnitWasPlannedMetadata = new Dictionary<string, object>
+            var firstBuildingUnitWasPlannedMetadata = new Dictionary<string, object>
             {
-                { AddEventHashPipe.HashMetadataKey, buildingUnitWasPlannedV2.GetHash() },
+                { AddEventHashPipe.HashMetadataKey, firstBuildingUnitWasPlannedV2.GetHash() },
                 { Envelope.PositionMetadataKey, position }
+            };
+            var secondBuildingUnitWasPlannedMetadata = new Dictionary<string, object>
+            {
+                { AddEventHashPipe.HashMetadataKey, firstBuildingUnitWasPlannedV2.GetHash() },
+                { Envelope.PositionMetadataKey, position + 1 }
             };
             var buildingMeasurementWasCorrectedMetadata = new Dictionary<string, object>
             {
                 { AddEventHashPipe.HashMetadataKey, buildingMeasurementWasCorrected.GetHash() },
-                { Envelope.PositionMetadataKey, position + 1 }
+                { Envelope.PositionMetadataKey, position + 2 }
             };
 
             await Sut
                 .Given(
-                    new Envelope<BuildingUnitWasPlannedV2>(new Envelope(buildingUnitWasPlannedV2, buildingUnitWasPlannedMetadata)),
+                    new Envelope<BuildingUnitWasPlannedV2>(new Envelope(firstBuildingUnitWasPlannedV2, firstBuildingUnitWasPlannedMetadata)),
+                    new Envelope<BuildingUnitWasPlannedV2>(new Envelope(secondBuildingUnitWasPlannedV2, secondBuildingUnitWasPlannedMetadata)),
                     new Envelope<BuildingMeasurementWasCorrected>(new Envelope(buildingMeasurementWasCorrected, buildingMeasurementWasCorrectedMetadata)))
                 .Then(async context =>
                 {
-                    var buildingUnitLatestItem =
-                        await context.BuildingUnitLatestItems.FindAsync(buildingUnitWasPlannedV2.BuildingUnitPersistentLocalId);
-                    buildingUnitLatestItem.Should().NotBeNull();
+                    var firstBuildingUnitVersion =
+                        await context.BuildingUnitVersions.FindAsync(position + 2, firstBuildingUnitWasPlannedV2.BuildingUnitPersistentLocalId);
+                    firstBuildingUnitVersion.Should().NotBeNull();
 
-                    buildingUnitLatestItem!.OsloGeometryMethod.Should().Be(BuildingUnitPositionGeometryMethod.DerivedFromObject.Map());
-                    buildingUnitLatestItem.GeometryMethod.Should().Be("DerivedFromObject");
-                    buildingUnitLatestItem.Geometry.Should()
-                        .BeEquivalentTo(_wkbReader.Read(buildingMeasurementWasCorrected.ExtendedWkbGeometryBuildingUnits!.ToByteArray()));
-                    buildingUnitLatestItem.VersionTimestamp.Should().Be(buildingMeasurementWasCorrected.Provenance.Timestamp);
+                    firstBuildingUnitVersion!.GeometryMethod.Should().Be(BuildingUnitPositionGeometryMethod.DerivedFromObject.GeometryMethod);
+                    firstBuildingUnitVersion.OsloGeometryMethod.Should().Be(BuildingUnitPositionGeometryMethod.DerivedFromObject.Map());
+                    firstBuildingUnitVersion.Geometry.Should().BeEquivalentTo(
+                        _wkbReader.Read(buildingMeasurementWasCorrected.ExtendedWkbGeometryBuildingUnits!.ToByteArray()));
+                    firstBuildingUnitVersion.VersionTimestamp.Should().Be(buildingMeasurementWasCorrected.Provenance.Timestamp);
+
+                    var secondBuildingUnitVersion =
+                        await context.BuildingUnitVersions.FindAsync(position + 2, secondBuildingUnitWasPlannedV2.BuildingUnitPersistentLocalId);
+                    secondBuildingUnitVersion.Should().NotBeNull();
+
+                    secondBuildingUnitVersion!.GeometryMethod.Should().Be(BuildingUnitPositionGeometryMethod.DerivedFromObject.GeometryMethod);
+                    secondBuildingUnitVersion.OsloGeometryMethod.Should().Be(BuildingUnitPositionGeometryMethod.DerivedFromObject.Map());
+                    secondBuildingUnitVersion.Geometry.Should().BeEquivalentTo(
+                        _wkbReader.Read(buildingMeasurementWasCorrected.ExtendedWkbGeometryBuildingUnits!.ToByteArray()));
+                    secondBuildingUnitVersion.VersionTimestamp.Should().Be(buildingMeasurementWasCorrected.Provenance.Timestamp);
                 });
         }
 
@@ -227,39 +281,61 @@ namespace BuildingRegistry.Tests.ProjectionTests.Integration
         public async Task WhenBuildingMeasurementWasChanged()
         {
             _fixture.Customize(new WithFixedBuildingPersistentLocalId());
-            _fixture.Customize(new WithFixedBuildingUnitPersistentLocalId());
 
-            var buildingUnitWasPlannedV2 = _fixture.Create<BuildingUnitWasPlannedV2>();
-            var buildingMeasurementWasChanged = _fixture.Create<BuildingMeasurementWasChanged>();
+            var firstBuildingUnitWasPlannedV2 = _fixture.Create<BuildingUnitWasPlannedV2>();
+            var secondBuildingUnitWasPlannedV2 = _fixture.Create<BuildingUnitWasPlannedV2>();
+            var buildingMeasurementWasChanged = new BuildingMeasurementWasChanged(
+                _fixture.Create<BuildingPersistentLocalId>(),
+                new []{ new BuildingUnitPersistentLocalId(firstBuildingUnitWasPlannedV2.BuildingUnitPersistentLocalId) },
+                new []{ new BuildingUnitPersistentLocalId(secondBuildingUnitWasPlannedV2.BuildingUnitPersistentLocalId) },
+                _fixture.Create<ExtendedWkbGeometry>(),
+                _fixture.Create<ExtendedWkbGeometry>());
+            ((ISetProvenance)buildingMeasurementWasChanged).SetProvenance(_fixture.Create<Provenance>());
 
             var position = _fixture.Create<long>();
 
-            var buildingUnitWasPlannedMetadata = new Dictionary<string, object>
+            var firstBuildingUnitWasPlannedMetadata = new Dictionary<string, object>
             {
-                { AddEventHashPipe.HashMetadataKey, buildingUnitWasPlannedV2.GetHash() },
+                { AddEventHashPipe.HashMetadataKey, firstBuildingUnitWasPlannedV2.GetHash() },
                 { Envelope.PositionMetadataKey, position }
+            };
+            var secondBuildingUnitWasPlannedMetadata = new Dictionary<string, object>
+            {
+                { AddEventHashPipe.HashMetadataKey, firstBuildingUnitWasPlannedV2.GetHash() },
+                { Envelope.PositionMetadataKey, position + 1 }
             };
             var buildingMeasurementWasChangedMetadata = new Dictionary<string, object>
             {
                 { AddEventHashPipe.HashMetadataKey, buildingMeasurementWasChanged.GetHash() },
-                { Envelope.PositionMetadataKey, position + 1 }
+                { Envelope.PositionMetadataKey, position + 2 }
             };
 
             await Sut
                 .Given(
-                    new Envelope<BuildingUnitWasPlannedV2>(new Envelope(buildingUnitWasPlannedV2, buildingUnitWasPlannedMetadata)),
+                    new Envelope<BuildingUnitWasPlannedV2>(new Envelope(firstBuildingUnitWasPlannedV2, firstBuildingUnitWasPlannedMetadata)),
+                    new Envelope<BuildingUnitWasPlannedV2>(new Envelope(secondBuildingUnitWasPlannedV2, secondBuildingUnitWasPlannedMetadata)),
                     new Envelope<BuildingMeasurementWasChanged>(new Envelope(buildingMeasurementWasChanged, buildingMeasurementWasChangedMetadata)))
                 .Then(async context =>
                 {
-                    var buildingUnitLatestItem = await context.BuildingUnitLatestItems.FindAsync(buildingUnitWasPlannedV2
-                        .BuildingUnitPersistentLocalId);
-                    buildingUnitLatestItem.Should().NotBeNull();
+                    var firstBuildingUnitVersion =
+                        await context.BuildingUnitVersions.FindAsync(position + 2, firstBuildingUnitWasPlannedV2.BuildingUnitPersistentLocalId);
+                    firstBuildingUnitVersion.Should().NotBeNull();
 
-                    buildingUnitLatestItem!.OsloGeometryMethod.Should().Be(BuildingUnitPositionGeometryMethod.DerivedFromObject.Map());
-                    buildingUnitLatestItem.GeometryMethod.Should().Be("DerivedFromObject");
-                    buildingUnitLatestItem.Geometry.Should().BeEquivalentTo(
+                    firstBuildingUnitVersion!.GeometryMethod.Should().Be(BuildingUnitPositionGeometryMethod.DerivedFromObject.GeometryMethod);
+                    firstBuildingUnitVersion.OsloGeometryMethod.Should().Be(BuildingUnitPositionGeometryMethod.DerivedFromObject.Map());
+                    firstBuildingUnitVersion.Geometry.Should().BeEquivalentTo(
                         _wkbReader.Read(buildingMeasurementWasChanged.ExtendedWkbGeometryBuildingUnits!.ToByteArray()));
-                    buildingUnitLatestItem.VersionTimestamp.Should().Be(buildingMeasurementWasChanged.Provenance.Timestamp);
+                    firstBuildingUnitVersion.VersionTimestamp.Should().Be(buildingMeasurementWasChanged.Provenance.Timestamp);
+
+                    var secondBuildingUnitVersion =
+                        await context.BuildingUnitVersions.FindAsync(position + 2, secondBuildingUnitWasPlannedV2.BuildingUnitPersistentLocalId);
+                    secondBuildingUnitVersion.Should().NotBeNull();
+
+                    secondBuildingUnitVersion!.GeometryMethod.Should().Be(BuildingUnitPositionGeometryMethod.DerivedFromObject.GeometryMethod);
+                    secondBuildingUnitVersion.OsloGeometryMethod.Should().Be(BuildingUnitPositionGeometryMethod.DerivedFromObject.Map());
+                    secondBuildingUnitVersion.Geometry.Should().BeEquivalentTo(
+                        _wkbReader.Read(buildingMeasurementWasChanged.ExtendedWkbGeometryBuildingUnits!.ToByteArray()));
+                    secondBuildingUnitVersion.VersionTimestamp.Should().Be(buildingMeasurementWasChanged.Provenance.Timestamp);
                 });
         }
 
@@ -282,21 +358,23 @@ namespace BuildingRegistry.Tests.ProjectionTests.Integration
                 .Given(new Envelope<BuildingUnitWasPlannedV2>(new Envelope(buildingUnitWasPlannedV2, metadata)))
                 .Then(async context =>
                 {
-                    var buildingUnitLatestItem = await context.BuildingUnitLatestItems.FindAsync(buildingUnitWasPlannedV2.BuildingUnitPersistentLocalId);
-                    buildingUnitLatestItem.Should().NotBeNull();
+                    var buildingUnitVersion = await context.BuildingUnitVersions.FindAsync(position, buildingUnitWasPlannedV2.BuildingUnitPersistentLocalId);
+                    buildingUnitVersion.Should().NotBeNull();
 
-                    buildingUnitLatestItem!.BuildingPersistentLocalId.Should().Be(buildingUnitWasPlannedV2.BuildingPersistentLocalId);
-                    buildingUnitLatestItem.Geometry.Should().BeEquivalentTo(
-                        _wkbReader.Read(buildingUnitWasPlannedV2.ExtendedWkbGeometry.ToByteArray()));
-                    buildingUnitLatestItem.OsloGeometryMethod.Should().Be(BuildingUnitPositionGeometryMethod.Parse(buildingUnitWasPlannedV2.GeometryMethod).Map());
-                    buildingUnitLatestItem.GeometryMethod.Should().Be(buildingUnitWasPlannedV2.GeometryMethod);
-                    buildingUnitLatestItem.OsloFunction.Should().Be(BuildingUnitFunction.Parse(buildingUnitWasPlannedV2.Function).Map());
-                    buildingUnitLatestItem.Function.Should().Be(buildingUnitWasPlannedV2.Function);
-                    buildingUnitLatestItem.OsloStatus.Should().Be(BuildingUnitStatus.Planned.Map());
-                    buildingUnitLatestItem.Status.Should().Be("Planned");
-                    buildingUnitLatestItem.HasDeviation.Should().Be(buildingUnitWasPlannedV2.HasDeviation);
-                    buildingUnitLatestItem.IsRemoved.Should().BeFalse();
-                    buildingUnitLatestItem.VersionTimestamp.Should().Be(buildingUnitWasPlannedV2.Provenance.Timestamp);
+                    buildingUnitVersion!.BuildingPersistentLocalId.Should().Be(buildingUnitWasPlannedV2.BuildingPersistentLocalId);
+                    buildingUnitVersion.Geometry.Should().BeEquivalentTo(_wkbReader.Read(buildingUnitWasPlannedV2.ExtendedWkbGeometry.ToByteArray()));
+                    buildingUnitVersion.GeometryMethod.Should().Be(BuildingUnitPositionGeometryMethod.Parse(buildingUnitWasPlannedV2.GeometryMethod).GeometryMethod);
+                    buildingUnitVersion.OsloGeometryMethod.Should().Be(BuildingUnitPositionGeometryMethod.Parse(buildingUnitWasPlannedV2.GeometryMethod).Map());
+                    buildingUnitVersion.Function.Should().Be(BuildingUnitFunction.Parse(buildingUnitWasPlannedV2.Function).Function);
+                    buildingUnitVersion.OsloFunction.Should().Be(BuildingUnitFunction.Parse(buildingUnitWasPlannedV2.Function).Map());
+                    buildingUnitVersion.Status.Should().Be(BuildingUnitStatus.Planned.Status);
+                    buildingUnitVersion.OsloStatus.Should().Be(BuildingUnitStatus.Planned.Map());
+                    buildingUnitVersion.HasDeviation.Should().Be(buildingUnitWasPlannedV2.HasDeviation);
+                    buildingUnitVersion.IsRemoved.Should().BeFalse();
+                    buildingUnitVersion.VersionTimestamp.Should().Be(buildingUnitWasPlannedV2.Provenance.Timestamp);
+                    buildingUnitVersion.CreatedOnTimestamp.Should().Be(buildingUnitWasPlannedV2.Provenance.Timestamp);
+                    buildingUnitVersion.Namespace.Should().Be(BuildingUnitNamespace);
+                    buildingUnitVersion.PuriId.Should().Be($"{BuildingUnitNamespace}/{buildingUnitWasPlannedV2.BuildingUnitPersistentLocalId}");
                 });
         }
 
@@ -328,13 +406,13 @@ namespace BuildingRegistry.Tests.ProjectionTests.Integration
                     new Envelope<BuildingUnitWasRealizedV2>(new Envelope(buildingUnitWasRealizedV2, buildingUnitWasRealizedMetadata)))
                 .Then(async context =>
                 {
-                    var buildingUnitLatestItem =
-                        await context.BuildingUnitLatestItems.FindAsync(buildingUnitWasRealizedV2.BuildingUnitPersistentLocalId);
-                    buildingUnitLatestItem.Should().NotBeNull();
+                    var buildingUnitVersion =
+                        await context.BuildingUnitVersions.FindAsync(position + 1, buildingUnitWasRealizedV2.BuildingUnitPersistentLocalId);
+                    buildingUnitVersion.Should().NotBeNull();
 
-                    buildingUnitLatestItem!.OsloStatus.Should().Be(BuildingUnitStatus.Realized.Map());
-                    buildingUnitLatestItem.Status.Should().Be("Realized");
-                    buildingUnitLatestItem.VersionTimestamp.Should().Be(buildingUnitWasRealizedV2.Provenance.Timestamp);
+                    buildingUnitVersion!.Status.Should().Be(BuildingUnitStatus.Realized.Status);
+                    buildingUnitVersion.OsloStatus.Should().Be(BuildingUnitStatus.Realized.Map());
+                    buildingUnitVersion.VersionTimestamp.Should().Be(buildingUnitWasRealizedV2.Provenance.Timestamp);
                 });
         }
 
@@ -366,13 +444,13 @@ namespace BuildingRegistry.Tests.ProjectionTests.Integration
                     new Envelope<BuildingUnitWasRealizedBecauseBuildingWasRealized>(new Envelope(buildingUnitWasRealized, buildingUnitWasRealizedMetadata)))
                 .Then(async context =>
                 {
-                    var buildingUnitLatestItem =
-                        await context.BuildingUnitLatestItems.FindAsync(buildingUnitWasRealized.BuildingUnitPersistentLocalId);
-                    buildingUnitLatestItem.Should().NotBeNull();
+                    var buildingUnitVersion =
+                        await context.BuildingUnitVersions.FindAsync(position + 1, buildingUnitWasRealized.BuildingUnitPersistentLocalId);
+                    buildingUnitVersion.Should().NotBeNull();
 
-                    buildingUnitLatestItem!.OsloStatus.Should().Be(BuildingUnitStatus.Realized.Map());
-                    buildingUnitLatestItem.Status.Should().Be("Realized");
-                    buildingUnitLatestItem.VersionTimestamp.Should().Be(buildingUnitWasRealized.Provenance.Timestamp);
+                    buildingUnitVersion!.Status.Should().Be(BuildingUnitStatus.Realized.Status);
+                    buildingUnitVersion.OsloStatus.Should().Be(BuildingUnitStatus.Realized.Map());
+                    buildingUnitVersion.VersionTimestamp.Should().Be(buildingUnitWasRealized.Provenance.Timestamp);
                 });
         }
 
@@ -412,13 +490,13 @@ namespace BuildingRegistry.Tests.ProjectionTests.Integration
                         new Envelope(buildingUnitWasCorrectedFromRealizedToPlanned, buildingUnitWasCorrectedToPlannedMetadata)))
                 .Then(async context =>
                 {
-                    var buildingUnitLatestItem =
-                        await context.BuildingUnitLatestItems.FindAsync(buildingUnitWasCorrectedFromRealizedToPlanned.BuildingUnitPersistentLocalId);
-                    buildingUnitLatestItem.Should().NotBeNull();
+                    var buildingUnitVersion =
+                        await context.BuildingUnitVersions.FindAsync(position + 2, buildingUnitWasCorrectedFromRealizedToPlanned.BuildingUnitPersistentLocalId);
+                    buildingUnitVersion.Should().NotBeNull();
 
-                    buildingUnitLatestItem!.OsloStatus.Should().Be(BuildingUnitStatus.Planned.Map());
-                    buildingUnitLatestItem.Status.Should().Be("Planned");
-                    buildingUnitLatestItem.VersionTimestamp.Should().Be(buildingUnitWasCorrectedFromRealizedToPlanned.Provenance.Timestamp);
+                    buildingUnitVersion!.Status.Should().Be(BuildingUnitStatus.Planned.Status);
+                    buildingUnitVersion.OsloStatus.Should().Be(BuildingUnitStatus.Planned.Map());
+                    buildingUnitVersion.VersionTimestamp.Should().Be(buildingUnitWasCorrectedFromRealizedToPlanned.Provenance.Timestamp);
                 });
         }
 
@@ -458,13 +536,13 @@ namespace BuildingRegistry.Tests.ProjectionTests.Integration
                         new Envelope(buildingUnitWasCorrectedToPlanned, buildingUnitWasCorrectedToPlannedMetadata)))
                 .Then(async context =>
                 {
-                    var buildingUnitLatestItem =
-                        await context.BuildingUnitLatestItems.FindAsync(buildingUnitWasCorrectedToPlanned.BuildingUnitPersistentLocalId);
-                    buildingUnitLatestItem.Should().NotBeNull();
+                    var buildingUnitVersion =
+                        await context.BuildingUnitVersions.FindAsync(position + 2, buildingUnitWasCorrectedToPlanned.BuildingUnitPersistentLocalId);
+                    buildingUnitVersion.Should().NotBeNull();
 
-                    buildingUnitLatestItem!.OsloStatus.Should().Be(BuildingUnitStatus.Planned.Map());
-                    buildingUnitLatestItem.Status.Should().Be("Planned");
-                    buildingUnitLatestItem.VersionTimestamp.Should().Be(buildingUnitWasCorrectedToPlanned.Provenance.Timestamp);
+                    buildingUnitVersion!.Status.Should().Be(BuildingUnitStatus.Planned.Status);
+                    buildingUnitVersion.OsloStatus.Should().Be(BuildingUnitStatus.Planned.Map());
+                    buildingUnitVersion.VersionTimestamp.Should().Be(buildingUnitWasCorrectedToPlanned.Provenance.Timestamp);
                 });
         }
 
@@ -496,13 +574,13 @@ namespace BuildingRegistry.Tests.ProjectionTests.Integration
                     new Envelope<BuildingUnitWasNotRealizedV2>(new Envelope(buildingUnitWasNotRealizedV2, buildingUnitWasNotRealizedMetadata)))
                 .Then(async context =>
                 {
-                    var buildingUnitLatestItem =
-                        await context.BuildingUnitLatestItems.FindAsync(buildingUnitWasNotRealizedV2.BuildingUnitPersistentLocalId);
-                    buildingUnitLatestItem.Should().NotBeNull();
+                    var buildingUnitVersion =
+                        await context.BuildingUnitVersions.FindAsync(position + 1, buildingUnitWasNotRealizedV2.BuildingUnitPersistentLocalId);
+                    buildingUnitVersion.Should().NotBeNull();
 
-                    buildingUnitLatestItem!.OsloStatus.Should().Be(BuildingUnitStatus.NotRealized.Map());
-                    buildingUnitLatestItem.Status.Should().Be("NotRealized");
-                    buildingUnitLatestItem.VersionTimestamp.Should().Be(buildingUnitWasNotRealizedV2.Provenance.Timestamp);
+                    buildingUnitVersion!.Status.Should().Be(BuildingUnitStatus.NotRealized.Status);
+                    buildingUnitVersion.OsloStatus.Should().Be(BuildingUnitStatus.NotRealized.Map());
+                    buildingUnitVersion.VersionTimestamp.Should().Be(buildingUnitWasNotRealizedV2.Provenance.Timestamp);
                 });
         }
 
@@ -535,13 +613,13 @@ namespace BuildingRegistry.Tests.ProjectionTests.Integration
                         new Envelope(buildingUnitWasNotRealized, buildingUnitWasNotRealizedMetadata)))
                 .Then(async context =>
                 {
-                    var buildingUnitLatestItem =
-                        await context.BuildingUnitLatestItems.FindAsync(buildingUnitWasNotRealized.BuildingUnitPersistentLocalId);
-                    buildingUnitLatestItem.Should().NotBeNull();
+                    var buildingUnitVersion =
+                        await context.BuildingUnitVersions.FindAsync(position + 1, buildingUnitWasNotRealized.BuildingUnitPersistentLocalId);
+                    buildingUnitVersion.Should().NotBeNull();
 
-                    buildingUnitLatestItem!.OsloStatus.Should().Be(BuildingUnitStatus.NotRealized.Map());
-                    buildingUnitLatestItem.Status.Should().Be("NotRealized");
-                    buildingUnitLatestItem.VersionTimestamp.Should().Be(buildingUnitWasNotRealized.Provenance.Timestamp);
+                    buildingUnitVersion!.Status.Should().Be(BuildingUnitStatus.NotRealized.Status);
+                    buildingUnitVersion.OsloStatus.Should().Be(BuildingUnitStatus.NotRealized.Map());
+                    buildingUnitVersion.VersionTimestamp.Should().Be(buildingUnitWasNotRealized.Provenance.Timestamp);
                 });
         }
 
@@ -581,13 +659,13 @@ namespace BuildingRegistry.Tests.ProjectionTests.Integration
                         new Envelope(buildingUnitWasCorrectedToPlanned, buildingUnitWasCorrectedToPlannedMetadata)))
                 .Then(async context =>
                 {
-                    var buildingUnitLatestItem =
-                        await context.BuildingUnitLatestItems.FindAsync(buildingUnitWasCorrectedToPlanned.BuildingUnitPersistentLocalId);
-                    buildingUnitLatestItem.Should().NotBeNull();
+                    var buildingUnitVersion =
+                        await context.BuildingUnitVersions.FindAsync(position + 2, buildingUnitWasCorrectedToPlanned.BuildingUnitPersistentLocalId);
+                    buildingUnitVersion.Should().NotBeNull();
 
-                    buildingUnitLatestItem!.OsloStatus.Should().Be(BuildingUnitStatus.Planned.Map());
-                    buildingUnitLatestItem.Status.Should().Be("Planned");
-                    buildingUnitLatestItem.VersionTimestamp.Should().Be(buildingUnitWasCorrectedToPlanned.Provenance.Timestamp);
+                    buildingUnitVersion!.Status.Should().Be(BuildingUnitStatus.Planned.Status);
+                    buildingUnitVersion.OsloStatus.Should().Be(BuildingUnitStatus.Planned.Map());
+                    buildingUnitVersion.VersionTimestamp.Should().Be(buildingUnitWasCorrectedToPlanned.Provenance.Timestamp);
                 });
         }
 
@@ -626,13 +704,13 @@ namespace BuildingRegistry.Tests.ProjectionTests.Integration
                     new Envelope<BuildingUnitWasRetiredV2>(new Envelope(buildingUnitWasRetiredV2, buildingUnitWasRetiredMetadata)))
                 .Then(async context =>
                 {
-                    var buildingUnitLatestItem =
-                        await context.BuildingUnitLatestItems.FindAsync(buildingUnitWasRetiredV2.BuildingUnitPersistentLocalId);
-                    buildingUnitLatestItem.Should().NotBeNull();
+                    var buildingUnitVersion =
+                        await context.BuildingUnitVersions.FindAsync(position + 2, buildingUnitWasRetiredV2.BuildingUnitPersistentLocalId);
+                    buildingUnitVersion.Should().NotBeNull();
 
-                    buildingUnitLatestItem!.OsloStatus.Should().Be(BuildingUnitStatus.Retired.Map());
-                    buildingUnitLatestItem.Status.Should().Be("Retired");
-                    buildingUnitLatestItem.VersionTimestamp.Should().Be(buildingUnitWasRetiredV2.Provenance.Timestamp);
+                    buildingUnitVersion!.Status.Should().Be(BuildingUnitStatus.Retired.Status);
+                    buildingUnitVersion.OsloStatus.Should().Be(BuildingUnitStatus.Retired.Map());
+                    buildingUnitVersion.VersionTimestamp.Should().Be(buildingUnitWasRetiredV2.Provenance.Timestamp);
                 });
         }
 
@@ -644,6 +722,7 @@ namespace BuildingRegistry.Tests.ProjectionTests.Integration
 
             var buildingUnitWasPlannedV2 = _fixture.Create<BuildingUnitWasPlannedV2>();
             var buildingUnitWasRetiredV2 = _fixture.Create<BuildingUnitWasRetiredV2>();
+            var buildingUnitWasRealizedV2 = _fixture.Create<BuildingUnitWasRealizedV2>();
             var buildingUnitWasCorrectedToRealized = _fixture.Create<BuildingUnitWasCorrectedFromRetiredToRealized>();
 
             var position = _fixture.Create<long>();
@@ -653,32 +732,38 @@ namespace BuildingRegistry.Tests.ProjectionTests.Integration
                 { AddEventHashPipe.HashMetadataKey, buildingUnitWasPlannedV2.GetHash() },
                 { Envelope.PositionMetadataKey, position }
             };
+            var buildingUnitWasRealizedMetadata = new Dictionary<string, object>
+            {
+                { AddEventHashPipe.HashMetadataKey, buildingUnitWasPlannedV2.GetHash() },
+                { Envelope.PositionMetadataKey, position + 1 }
+            };
             var buildingUnitWasRetiredMetadata = new Dictionary<string, object>
             {
                 { AddEventHashPipe.HashMetadataKey, buildingUnitWasRetiredV2.GetHash() },
-                { Envelope.PositionMetadataKey, position + 1 }
+                { Envelope.PositionMetadataKey, position + 2 }
             };
             var buildingUnitWasCorrectedToRealizedMetadata = new Dictionary<string, object>
             {
                 { AddEventHashPipe.HashMetadataKey, buildingUnitWasCorrectedToRealized.GetHash() },
-                { Envelope.PositionMetadataKey, position + 2 }
+                { Envelope.PositionMetadataKey, position + 3 }
             };
 
             await Sut
                 .Given(
                     new Envelope<BuildingUnitWasPlannedV2>(new Envelope(buildingUnitWasPlannedV2, buildingUnitWasPlannedMetadata)),
+                    new Envelope<BuildingUnitWasRealizedV2>(new Envelope(buildingUnitWasRealizedV2, buildingUnitWasRealizedMetadata)),
                     new Envelope<BuildingUnitWasRetiredV2>(new Envelope(buildingUnitWasRetiredV2, buildingUnitWasRetiredMetadata)),
                     new Envelope<BuildingUnitWasCorrectedFromRetiredToRealized>(
                         new Envelope(buildingUnitWasCorrectedToRealized, buildingUnitWasCorrectedToRealizedMetadata)))
                 .Then(async context =>
                 {
-                    var buildingUnitLatestItem =
-                        await context.BuildingUnitLatestItems.FindAsync(buildingUnitWasCorrectedToRealized.BuildingUnitPersistentLocalId);
-                    buildingUnitLatestItem.Should().NotBeNull();
+                    var buildingUnitVersion =
+                        await context.BuildingUnitVersions.FindAsync(position + 3, buildingUnitWasCorrectedToRealized.BuildingUnitPersistentLocalId);
+                    buildingUnitVersion.Should().NotBeNull();
 
-                    buildingUnitLatestItem!.OsloStatus.Should().Be(BuildingUnitStatus.Realized.Map());
-                    buildingUnitLatestItem.Status.Should().Be("Realized");
-                    buildingUnitLatestItem.VersionTimestamp.Should().Be(buildingUnitWasCorrectedToRealized.Provenance.Timestamp);
+                    buildingUnitVersion!.Status.Should().Be(BuildingUnitStatus.Realized.Status);
+                    buildingUnitVersion.OsloStatus.Should().Be(BuildingUnitStatus.Realized.Map());
+                    buildingUnitVersion.VersionTimestamp.Should().Be(buildingUnitWasCorrectedToRealized.Provenance.Timestamp);
                 });
         }
 
@@ -709,16 +794,16 @@ namespace BuildingRegistry.Tests.ProjectionTests.Integration
                     new Envelope<BuildingUnitPositionWasCorrected>(new Envelope(buildingUnitPositionWasCorrected, buildingUnitPositionWasCorrectedMetadata)))
                 .Then(async context =>
                 {
-                    var buildingUnitLatestItem =
-                        await context.BuildingUnitLatestItems.FindAsync(buildingUnitPositionWasCorrected.BuildingUnitPersistentLocalId);
-                    buildingUnitLatestItem.Should().NotBeNull();
+                    var buildingUnitVersion =
+                        await context.BuildingUnitVersions.FindAsync(position + 1, buildingUnitPositionWasCorrected.BuildingUnitPersistentLocalId);
+                    buildingUnitVersion.Should().NotBeNull();
 
-                    buildingUnitLatestItem!.BuildingPersistentLocalId.Should().Be(buildingUnitPositionWasCorrected.BuildingPersistentLocalId);
-                    buildingUnitLatestItem.Geometry.Should().BeEquivalentTo(
+                    buildingUnitVersion!.BuildingPersistentLocalId.Should().Be(buildingUnitPositionWasCorrected.BuildingPersistentLocalId);
+                    buildingUnitVersion.Geometry.Should().BeEquivalentTo(
                         _wkbReader.Read(buildingUnitPositionWasCorrected.ExtendedWkbGeometry.ToByteArray()));
-                    buildingUnitLatestItem.OsloGeometryMethod.Should().Be(BuildingUnitPositionGeometryMethod.Parse(buildingUnitPositionWasCorrected.GeometryMethod).Map());
-                    buildingUnitLatestItem.GeometryMethod.Should().Be(buildingUnitPositionWasCorrected.GeometryMethod);
-                    buildingUnitLatestItem.VersionTimestamp.Should().Be(buildingUnitPositionWasCorrected.Provenance.Timestamp);
+                    buildingUnitVersion.GeometryMethod.Should().Be(BuildingUnitPositionGeometryMethod.Parse(buildingUnitPositionWasCorrected.GeometryMethod).GeometryMethod);
+                    buildingUnitVersion.OsloGeometryMethod.Should().Be(BuildingUnitPositionGeometryMethod.Parse(buildingUnitPositionWasCorrected.GeometryMethod).Map());
+                    buildingUnitVersion.VersionTimestamp.Should().Be(buildingUnitPositionWasCorrected.Provenance.Timestamp);
                 });
         }
 
@@ -750,12 +835,12 @@ namespace BuildingRegistry.Tests.ProjectionTests.Integration
                     new Envelope<BuildingUnitWasRemovedV2>(new Envelope(buildingUnitWasRemovedV2, buildingUnitWasRemovedMetadata)))
                 .Then(async context =>
                 {
-                    var buildingUnitLatestItem =
-                        await context.BuildingUnitLatestItems.FindAsync(buildingUnitWasRemovedV2.BuildingUnitPersistentLocalId);
-                    buildingUnitLatestItem.Should().NotBeNull();
+                    var buildingUnitVersion =
+                        await context.BuildingUnitVersions.FindAsync(position + 1, buildingUnitWasRemovedV2.BuildingUnitPersistentLocalId);
+                    buildingUnitVersion.Should().NotBeNull();
 
-                    buildingUnitLatestItem!.IsRemoved.Should().BeTrue();
-                    buildingUnitLatestItem.VersionTimestamp.Should().Be(buildingUnitWasRemovedV2.Provenance.Timestamp);
+                    buildingUnitVersion!.IsRemoved.Should().BeTrue();
+                    buildingUnitVersion.VersionTimestamp.Should().Be(buildingUnitWasRemovedV2.Provenance.Timestamp);
                 });
         }
 
@@ -787,12 +872,12 @@ namespace BuildingRegistry.Tests.ProjectionTests.Integration
                     new Envelope<BuildingUnitWasRemovedBecauseBuildingWasRemoved>(new Envelope(buildingUnitWasRemoved, buildingUnitWasRemovedMetadata)))
                 .Then(async context =>
                 {
-                    var buildingUnitLatestItem =
-                        await context.BuildingUnitLatestItems.FindAsync(buildingUnitWasRemoved.BuildingUnitPersistentLocalId);
-                    buildingUnitLatestItem.Should().NotBeNull();
+                    var buildingUnitVersion =
+                        await context.BuildingUnitVersions.FindAsync(position + 1, buildingUnitWasRemoved.BuildingUnitPersistentLocalId);
+                    buildingUnitVersion.Should().NotBeNull();
 
-                    buildingUnitLatestItem!.IsRemoved.Should().BeTrue();
-                    buildingUnitLatestItem.VersionTimestamp.Should().Be(buildingUnitWasRemoved.Provenance.Timestamp);
+                    buildingUnitVersion!.IsRemoved.Should().BeTrue();
+                    buildingUnitVersion.VersionTimestamp.Should().Be(buildingUnitWasRemoved.Provenance.Timestamp);
                 });
         }
 
@@ -831,21 +916,21 @@ namespace BuildingRegistry.Tests.ProjectionTests.Integration
                     new Envelope<BuildingUnitRemovalWasCorrected>(new Envelope(buildingUnitRemovalWasCorrected, buildingUnitRemovalWasCorrectedMetadata)))
                 .Then(async context =>
                 {
-                    var buildingUnitLatestItem =
-                        await context.BuildingUnitLatestItems.FindAsync(buildingUnitRemovalWasCorrected.BuildingUnitPersistentLocalId);
-                    buildingUnitLatestItem.Should().NotBeNull();
+                    var buildingUnitVersion =
+                        await context.BuildingUnitVersions.FindAsync(position + 2, buildingUnitRemovalWasCorrected.BuildingUnitPersistentLocalId);
+                    buildingUnitVersion.Should().NotBeNull();
 
-                    buildingUnitLatestItem!.OsloStatus.Should().Be(BuildingUnitStatus.Parse(buildingUnitRemovalWasCorrected.BuildingUnitStatus).Map());
-                    buildingUnitLatestItem.Status.Should().Be(buildingUnitRemovalWasCorrected.BuildingUnitStatus);
-                    buildingUnitLatestItem.OsloFunction.Should().Be(BuildingUnitFunction.Parse(buildingUnitRemovalWasCorrected.Function).Map());
-                    buildingUnitLatestItem.Function.Should().Be(buildingUnitRemovalWasCorrected.Function);
-                    buildingUnitLatestItem.Geometry.Should().BeEquivalentTo(
+                    buildingUnitVersion!.Status.Should().Be(BuildingUnitStatus.Parse(buildingUnitRemovalWasCorrected.BuildingUnitStatus).Status);
+                    buildingUnitVersion.OsloStatus.Should().Be(BuildingUnitStatus.Parse(buildingUnitRemovalWasCorrected.BuildingUnitStatus).Map());
+                    buildingUnitVersion.Function.Should().Be(BuildingUnitFunction.Parse(buildingUnitRemovalWasCorrected.Function).Function);
+                    buildingUnitVersion.OsloFunction.Should().Be(BuildingUnitFunction.Parse(buildingUnitRemovalWasCorrected.Function).Map());
+                    buildingUnitVersion.Geometry.Should().BeEquivalentTo(
                         _wkbReader.Read(buildingUnitRemovalWasCorrected.ExtendedWkbGeometry.ToByteArray()));
-                    buildingUnitLatestItem.OsloGeometryMethod.Should().Be(BuildingUnitPositionGeometryMethod.Parse(buildingUnitRemovalWasCorrected.GeometryMethod).Map());
-                    buildingUnitLatestItem.GeometryMethod.Should().Be(buildingUnitRemovalWasCorrected.GeometryMethod);
-                    buildingUnitLatestItem.HasDeviation.Should().Be(buildingUnitRemovalWasCorrected.HasDeviation);
-                    buildingUnitLatestItem.IsRemoved.Should().BeFalse();
-                    buildingUnitLatestItem.VersionTimestamp.Should().Be(buildingUnitRemovalWasCorrected.Provenance.Timestamp);
+                    buildingUnitVersion.GeometryMethod.Should().Be(BuildingUnitPositionGeometryMethod.Parse(buildingUnitRemovalWasCorrected.GeometryMethod).GeometryMethod);
+                    buildingUnitVersion.OsloGeometryMethod.Should().Be(BuildingUnitPositionGeometryMethod.Parse(buildingUnitRemovalWasCorrected.GeometryMethod).Map());
+                    buildingUnitVersion.HasDeviation.Should().Be(buildingUnitRemovalWasCorrected.HasDeviation);
+                    buildingUnitVersion.IsRemoved.Should().BeFalse();
+                    buildingUnitVersion.VersionTimestamp.Should().Be(buildingUnitRemovalWasCorrected.Provenance.Timestamp);
                 });
         }
 
@@ -877,12 +962,12 @@ namespace BuildingRegistry.Tests.ProjectionTests.Integration
                     new Envelope<BuildingUnitWasRegularized>(new Envelope(buildingUnitWasRegularized, buildingUnitWasRegularizedMetadata)))
                 .Then(async context =>
                 {
-                    var buildingUnitLatestItem =
-                        await context.BuildingUnitLatestItems.FindAsync(buildingUnitWasRegularized.BuildingUnitPersistentLocalId);
-                    buildingUnitLatestItem.Should().NotBeNull();
+                    var buildingUnitVersion =
+                        await context.BuildingUnitVersions.FindAsync(position + 1, buildingUnitWasRegularized.BuildingUnitPersistentLocalId);
+                    buildingUnitVersion.Should().NotBeNull();
 
-                    buildingUnitLatestItem!.HasDeviation.Should().BeFalse();
-                    buildingUnitLatestItem.VersionTimestamp.Should().Be(buildingUnitWasRegularized.Provenance.Timestamp);
+                    buildingUnitVersion!.HasDeviation.Should().BeFalse();
+                    buildingUnitVersion.VersionTimestamp.Should().Be(buildingUnitWasRegularized.Provenance.Timestamp);
                 });
         }
 
@@ -915,12 +1000,12 @@ namespace BuildingRegistry.Tests.ProjectionTests.Integration
                         new Envelope(buildingUnitRegularizationWasCorrected, buildingUnitRegularizationWasCorrectedMetadata)))
                 .Then(async context =>
                 {
-                    var buildingUnitLatestItem =
-                        await context.BuildingUnitLatestItems.FindAsync(buildingUnitRegularizationWasCorrected.BuildingUnitPersistentLocalId);
-                    buildingUnitLatestItem.Should().NotBeNull();
+                    var buildingUnitVersion =
+                        await context.BuildingUnitVersions.FindAsync(position + 1, buildingUnitRegularizationWasCorrected.BuildingUnitPersistentLocalId);
+                    buildingUnitVersion.Should().NotBeNull();
 
-                    buildingUnitLatestItem!.HasDeviation.Should().BeTrue();
-                    buildingUnitLatestItem.VersionTimestamp.Should().Be(buildingUnitRegularizationWasCorrected.Provenance.Timestamp);
+                    buildingUnitVersion!.HasDeviation.Should().BeTrue();
+                    buildingUnitVersion.VersionTimestamp.Should().Be(buildingUnitRegularizationWasCorrected.Provenance.Timestamp);
                 });
         }
 
@@ -952,12 +1037,12 @@ namespace BuildingRegistry.Tests.ProjectionTests.Integration
                     new Envelope<BuildingUnitWasDeregulated>(new Envelope(buildingUnitWasDeregulated, buildingUnitWasDeregulatedMetadata)))
                 .Then(async context =>
                 {
-                    var buildingUnitLatestItem =
-                        await context.BuildingUnitLatestItems.FindAsync(buildingUnitWasDeregulated.BuildingUnitPersistentLocalId);
-                    buildingUnitLatestItem.Should().NotBeNull();
+                    var buildingUnitVersion =
+                        await context.BuildingUnitVersions.FindAsync(position + 1, buildingUnitWasDeregulated.BuildingUnitPersistentLocalId);
+                    buildingUnitVersion.Should().NotBeNull();
 
-                    buildingUnitLatestItem!.HasDeviation.Should().BeTrue();
-                    buildingUnitLatestItem.VersionTimestamp.Should().Be(buildingUnitWasDeregulated.Provenance.Timestamp);
+                    buildingUnitVersion!.HasDeviation.Should().BeTrue();
+                    buildingUnitVersion.VersionTimestamp.Should().Be(buildingUnitWasDeregulated.Provenance.Timestamp);
                 });
         }
 
@@ -989,12 +1074,12 @@ namespace BuildingRegistry.Tests.ProjectionTests.Integration
                         new Envelope(buildingUnitDeregulationWasCorrected, buildingUnitDeregulationWasCorrectedMetadata)))
                 .Then(async context =>
                 {
-                    var buildingUnitLatestItem =
-                        await context.BuildingUnitLatestItems.FindAsync(buildingUnitDeregulationWasCorrected.BuildingUnitPersistentLocalId);
-                    buildingUnitLatestItem.Should().NotBeNull();
+                    var buildingUnitVersion =
+                        await context.BuildingUnitVersions.FindAsync(position + 1, buildingUnitDeregulationWasCorrected.BuildingUnitPersistentLocalId);
+                    buildingUnitVersion.Should().NotBeNull();
 
-                    buildingUnitLatestItem!.HasDeviation.Should().BeFalse();
-                    buildingUnitLatestItem.VersionTimestamp.Should().Be(buildingUnitDeregulationWasCorrected.Provenance.Timestamp);
+                    buildingUnitVersion!.HasDeviation.Should().BeFalse();
+                    buildingUnitVersion.VersionTimestamp.Should().Be(buildingUnitDeregulationWasCorrected.Provenance.Timestamp);
                 });
         }
 
@@ -1025,22 +1110,24 @@ namespace BuildingRegistry.Tests.ProjectionTests.Integration
                 .Given(new Envelope<CommonBuildingUnitWasAddedV2>(new Envelope(commonBuildingUnitWasAddedV2, commonBuildingUnitWasAddedMetadata)))
                 .Then(async context =>
                 {
-                    var buildingUnitLatestItem = await context.BuildingUnitLatestItems.FindAsync(commonBuildingUnitWasAddedV2
-                        .BuildingUnitPersistentLocalId);
-                    buildingUnitLatestItem.Should().NotBeNull();
+                    var buildingUnitVersion = await context.BuildingUnitVersions.FindAsync(position, commonBuildingUnitWasAddedV2.BuildingUnitPersistentLocalId);
+                    buildingUnitVersion.Should().NotBeNull();
 
-                    buildingUnitLatestItem!.BuildingPersistentLocalId.Should().Be(commonBuildingUnitWasAddedV2.BuildingPersistentLocalId);
-                    buildingUnitLatestItem.Geometry.Should().BeEquivalentTo(
+                    buildingUnitVersion!.BuildingPersistentLocalId.Should().Be(commonBuildingUnitWasAddedV2.BuildingPersistentLocalId);
+                    buildingUnitVersion.Geometry.Should().BeEquivalentTo(
                         _wkbReader.Read(commonBuildingUnitWasAddedV2.ExtendedWkbGeometry.ToByteArray()));
-                    buildingUnitLatestItem.OsloGeometryMethod.Should().Be(BuildingUnitPositionGeometryMethod.Parse(commonBuildingUnitWasAddedV2.GeometryMethod).Map());
-                    buildingUnitLatestItem.GeometryMethod.Should().Be(commonBuildingUnitWasAddedV2.GeometryMethod);
-                    buildingUnitLatestItem.OsloFunction.Should().Be(BuildingUnitFunction.Common.Map());
-                    buildingUnitLatestItem.Function.Should().Be(BuildingUnitFunction.Common.Function);
-                    buildingUnitLatestItem.OsloStatus.Should().Be(BuildingUnitStatus.Parse(commonBuildingUnitWasAddedV2.BuildingUnitStatus).Map());
-                    buildingUnitLatestItem.Status.Should().Be(commonBuildingUnitWasAddedV2.BuildingUnitStatus);
-                    buildingUnitLatestItem.HasDeviation.Should().Be(commonBuildingUnitWasAddedV2.HasDeviation);
-                    buildingUnitLatestItem.IsRemoved.Should().BeFalse();
-                    buildingUnitLatestItem.VersionTimestamp.Should().Be(commonBuildingUnitWasAddedV2.Provenance.Timestamp);
+                    buildingUnitVersion.GeometryMethod.Should().Be(BuildingUnitPositionGeometryMethod.Parse(commonBuildingUnitWasAddedV2.GeometryMethod).GeometryMethod);
+                    buildingUnitVersion.OsloGeometryMethod.Should().Be(BuildingUnitPositionGeometryMethod.Parse(commonBuildingUnitWasAddedV2.GeometryMethod).Map());
+                    buildingUnitVersion.Function.Should().Be(BuildingUnitFunction.Common.Function);
+                    buildingUnitVersion.OsloFunction.Should().Be(BuildingUnitFunction.Common.Map());
+                    buildingUnitVersion.Status.Should().Be(BuildingUnitStatus.Parse(commonBuildingUnitWasAddedV2.BuildingUnitStatus).Status);
+                    buildingUnitVersion.OsloStatus.Should().Be(BuildingUnitStatus.Parse(commonBuildingUnitWasAddedV2.BuildingUnitStatus).Map());
+                    buildingUnitVersion.HasDeviation.Should().Be(commonBuildingUnitWasAddedV2.HasDeviation);
+                    buildingUnitVersion.IsRemoved.Should().BeFalse();
+                    buildingUnitVersion.VersionTimestamp.Should().Be(commonBuildingUnitWasAddedV2.Provenance.Timestamp);
+                    buildingUnitVersion.CreatedOnTimestamp.Should().Be(commonBuildingUnitWasAddedV2.Provenance.Timestamp);
+                    buildingUnitVersion.Namespace.Should().Be(BuildingUnitNamespace);
+                    buildingUnitVersion.PuriId.Should().Be($"{BuildingUnitNamespace}/{commonBuildingUnitWasAddedV2.BuildingUnitPersistentLocalId}");
                 });
         }
 
@@ -1072,17 +1159,14 @@ namespace BuildingRegistry.Tests.ProjectionTests.Integration
                     new Envelope<BuildingUnitAddressWasAttachedV2>(new Envelope(buildingUnitAddressWasAttachedV2, buildingUnitAddressWasAttachedMetadata)))
                 .Then(async context =>
                 {
-                    var buildingUnitLatestItem =
-                        await context.BuildingUnitLatestItems.FindAsync(buildingUnitAddressWasAttachedV2.BuildingUnitPersistentLocalId);
-                    buildingUnitLatestItem.Should().NotBeNull();
+                    var buildingUnitVersion =
+                        await context.BuildingUnitVersions.FindAsync(position + 1, buildingUnitAddressWasAttachedV2.BuildingUnitPersistentLocalId);
+                    buildingUnitVersion.Should().NotBeNull();
 
-                    buildingUnitLatestItem!.VersionTimestamp.Should().Be(buildingUnitAddressWasAttachedV2.Provenance.Timestamp);
+                    buildingUnitVersion!.VersionTimestamp.Should().Be(buildingUnitAddressWasAttachedV2.Provenance.Timestamp);
 
-                    var buildingUnitAddresses = context.BuildingUnitAddresses
-                        .Where(x => x.BuildingUnitPersistentLocalId == buildingUnitLatestItem.BuildingUnitPersistentLocalId)
-                        .ToList();
-                    buildingUnitAddresses.Should().HaveCount(1);
-                    buildingUnitAddresses.Single().AddressPersistentLocalId.Should().Be(buildingUnitAddressWasAttachedV2.AddressPersistentLocalId);
+                    buildingUnitVersion.Addresses.Should().HaveCount(1);
+                    buildingUnitVersion.Addresses.Single().AddressPersistentLocalId.Should().Be(buildingUnitAddressWasAttachedV2.AddressPersistentLocalId);
                 });
         }
 
@@ -1122,16 +1206,13 @@ namespace BuildingRegistry.Tests.ProjectionTests.Integration
                     new Envelope<BuildingUnitAddressWasDetachedV2>(new Envelope(buildingUnitAddressWasDetachedV2, buildingUnitAddressWasDetachedMetadata)))
                 .Then(async context =>
                 {
-                    var buildingUnitLatestItem =
-                        await context.BuildingUnitLatestItems.FindAsync(buildingUnitAddressWasDetachedV2.BuildingUnitPersistentLocalId);
-                    buildingUnitLatestItem.Should().NotBeNull();
+                    var buildingUnitVersion =
+                        await context.BuildingUnitVersions.FindAsync(position + 2, buildingUnitAddressWasDetachedV2.BuildingUnitPersistentLocalId);
+                    buildingUnitVersion.Should().NotBeNull();
 
-                    buildingUnitLatestItem!.VersionTimestamp.Should().Be(buildingUnitAddressWasDetachedV2.Provenance.Timestamp);
+                    buildingUnitVersion!.VersionTimestamp.Should().Be(buildingUnitAddressWasDetachedV2.Provenance.Timestamp);
 
-                    var buildingUnitAddresses = context.BuildingUnitAddresses
-                        .Where(x => x.BuildingUnitPersistentLocalId == buildingUnitLatestItem.BuildingUnitPersistentLocalId)
-                        .ToList();
-                    buildingUnitAddresses.Should().BeEmpty();
+                    buildingUnitVersion.Addresses.Should().BeEmpty();
                 });
         }
 
@@ -1174,16 +1255,13 @@ namespace BuildingRegistry.Tests.ProjectionTests.Integration
                         new Envelope(buildingUnitAddressWasDetached, buildingUnitAddressWasDetachedMetadata)))
                 .Then(async context =>
                 {
-                    var buildingUnitLatestItem =
-                        await context.BuildingUnitLatestItems.FindAsync(buildingUnitAddressWasDetached.BuildingUnitPersistentLocalId);
-                    buildingUnitLatestItem.Should().NotBeNull();
+                    var buildingUnitVersion =
+                        await context.BuildingUnitVersions.FindAsync(position + 2, buildingUnitAddressWasDetached.BuildingUnitPersistentLocalId);
+                    buildingUnitVersion.Should().NotBeNull();
 
-                    buildingUnitLatestItem!.VersionTimestamp.Should().Be(buildingUnitAddressWasDetached.Provenance.Timestamp);
+                    buildingUnitVersion!.VersionTimestamp.Should().Be(buildingUnitAddressWasDetached.Provenance.Timestamp);
 
-                    var buildingUnitAddresses = context.BuildingUnitAddresses
-                        .Where(x => x.BuildingUnitPersistentLocalId == buildingUnitLatestItem.BuildingUnitPersistentLocalId)
-                        .ToList();
-                    buildingUnitAddresses.Should().BeEmpty();
+                    buildingUnitVersion.Addresses.Should().BeEmpty();
                 });
         }
 
@@ -1225,16 +1303,13 @@ namespace BuildingRegistry.Tests.ProjectionTests.Integration
                         new Envelope(buildingUnitAddressWasDetached, buildingUnitAddressWasDetachedMetadata)))
                 .Then(async context =>
                 {
-                    var buildingUnitLatestItem =
-                        await context.BuildingUnitLatestItems.FindAsync(buildingUnitAddressWasDetached.BuildingUnitPersistentLocalId);
-                    buildingUnitLatestItem.Should().NotBeNull();
+                    var buildingUnitVersion =
+                        await context.BuildingUnitVersions.FindAsync(position + 2, buildingUnitAddressWasDetached.BuildingUnitPersistentLocalId);
+                    buildingUnitVersion.Should().NotBeNull();
 
-                    buildingUnitLatestItem!.VersionTimestamp.Should().Be(buildingUnitAddressWasDetached.Provenance.Timestamp);
+                    buildingUnitVersion!.VersionTimestamp.Should().Be(buildingUnitAddressWasDetached.Provenance.Timestamp);
 
-                    var buildingUnitAddresses = context.BuildingUnitAddresses
-                        .Where(x => x.BuildingUnitPersistentLocalId == buildingUnitLatestItem.BuildingUnitPersistentLocalId)
-                        .ToList();
-                    buildingUnitAddresses.Should().BeEmpty();
+                    buildingUnitVersion.Addresses.Should().BeEmpty();
                 });
         }
 
@@ -1276,16 +1351,13 @@ namespace BuildingRegistry.Tests.ProjectionTests.Integration
                         new Envelope(buildingUnitAddressWasDetached, buildingUnitAddressWasDetachedMetadata)))
                 .Then(async context =>
                 {
-                    var buildingUnitLatestItem =
-                        await context.BuildingUnitLatestItems.FindAsync(buildingUnitAddressWasDetached.BuildingUnitPersistentLocalId);
-                    buildingUnitLatestItem.Should().NotBeNull();
+                    var buildingUnitVersion =
+                        await context.BuildingUnitVersions.FindAsync(position + 2, buildingUnitAddressWasDetached.BuildingUnitPersistentLocalId);
+                    buildingUnitVersion.Should().NotBeNull();
 
-                    buildingUnitLatestItem!.VersionTimestamp.Should().Be(buildingUnitAddressWasDetached.Provenance.Timestamp);
+                    buildingUnitVersion!.VersionTimestamp.Should().Be(buildingUnitAddressWasDetached.Provenance.Timestamp);
 
-                    var buildingUnitAddresses = context.BuildingUnitAddresses
-                        .Where(x => x.BuildingUnitPersistentLocalId == buildingUnitLatestItem.BuildingUnitPersistentLocalId)
-                        .ToList();
-                    buildingUnitAddresses.Should().BeEmpty();
+                    buildingUnitVersion.Addresses.Should().BeEmpty();
                 });
         }
 
@@ -1332,17 +1404,17 @@ namespace BuildingRegistry.Tests.ProjectionTests.Integration
                             buildingUnitAddressWasReplacedBecauseAddressWasReaddressedMetadata)))
                 .Then(async context =>
                 {
-                    var buildingUnitLatestItem =
-                        await context.BuildingUnitLatestItems.FindAsync(buildingUnitAddressWasReplacedBecauseAddressWasReaddressed.BuildingUnitPersistentLocalId);
-                    buildingUnitLatestItem.Should().NotBeNull();
+                    var buildingUnitVersion =
+                        await context.BuildingUnitVersions.FindAsync(position + 2, buildingUnitAddressWasReplacedBecauseAddressWasReaddressed.BuildingUnitPersistentLocalId);
+                    buildingUnitVersion.Should().NotBeNull();
 
-                    buildingUnitLatestItem!.VersionTimestamp.Should().Be(buildingUnitAddressWasReplacedBecauseAddressWasReaddressed.Provenance.Timestamp);
+                    buildingUnitVersion!.VersionTimestamp.Should().Be(buildingUnitAddressWasReplacedBecauseAddressWasReaddressed.Provenance.Timestamp);
 
-                    var newAddress = context.BuildingUnitAddresses.SingleOrDefault(x =>
+                    var newAddress = buildingUnitVersion.Addresses.SingleOrDefault(x =>
                         x.AddressPersistentLocalId == buildingUnitAddressWasReplacedBecauseAddressWasReaddressed.NewAddressPersistentLocalId);
                     newAddress.Should().NotBeNull();
 
-                    var oldAddress = context.BuildingUnitAddresses.SingleOrDefault(x =>
+                    var oldAddress = buildingUnitVersion.Addresses.SingleOrDefault(x =>
                         x.AddressPersistentLocalId == buildingUnitAddressWasReplacedBecauseAddressWasReaddressed.PreviousAddressPersistentLocalId);
                     oldAddress.Should().BeNull();
                 });
@@ -1376,13 +1448,13 @@ namespace BuildingRegistry.Tests.ProjectionTests.Integration
                     new Envelope<BuildingUnitWasRetiredBecauseBuildingWasDemolished>(new Envelope(buildingUnitWasRetired, buildingUnitWasRetiredMetadata)))
                 .Then(async context =>
                 {
-                    var buildingUnitLatestItem =
-                        await context.BuildingUnitLatestItems.FindAsync(buildingUnitWasRetired.BuildingUnitPersistentLocalId);
-                    buildingUnitLatestItem.Should().NotBeNull();
+                    var buildingUnitVersion =
+                        await context.BuildingUnitVersions.FindAsync(position + 1, buildingUnitWasRetired.BuildingUnitPersistentLocalId);
+                    buildingUnitVersion.Should().NotBeNull();
 
-                    buildingUnitLatestItem!.OsloStatus.Should().Be(BuildingUnitStatus.Retired.Map());
-                    buildingUnitLatestItem.Status.Should().Be("Retired");
-                    buildingUnitLatestItem.VersionTimestamp.Should().Be(buildingUnitWasRetired.Provenance.Timestamp);
+                    buildingUnitVersion!.Status.Should().Be(BuildingUnitStatus.Retired.Status);
+                    buildingUnitVersion.OsloStatus.Should().Be(BuildingUnitStatus.Retired.Map());
+                    buildingUnitVersion.VersionTimestamp.Should().Be(buildingUnitWasRetired.Provenance.Timestamp);
                 });
         }
 
@@ -1414,13 +1486,13 @@ namespace BuildingRegistry.Tests.ProjectionTests.Integration
                     new Envelope<BuildingUnitWasNotRealizedBecauseBuildingWasDemolished>(new Envelope(buildingUnitWasNotRealized, buildingUnitWasNotRealizedMetadata)))
                 .Then(async context =>
                 {
-                    var buildingUnitLatestItem =
-                        await context.BuildingUnitLatestItems.FindAsync(buildingUnitWasNotRealized.BuildingUnitPersistentLocalId);
-                    buildingUnitLatestItem.Should().NotBeNull();
+                    var buildingUnitVersion =
+                        await context.BuildingUnitVersions.FindAsync(position + 1, buildingUnitWasNotRealized.BuildingUnitPersistentLocalId);
+                    buildingUnitVersion.Should().NotBeNull();
 
-                    buildingUnitLatestItem!.OsloStatus.Should().Be(BuildingUnitStatus.NotRealized.Map());
-                    buildingUnitLatestItem.Status.Should().Be("NotRealized");
-                    buildingUnitLatestItem.VersionTimestamp.Should().Be(buildingUnitWasNotRealized.Provenance.Timestamp);
+                    buildingUnitVersion!.Status.Should().Be(BuildingUnitStatus.NotRealized.Status);
+                    buildingUnitVersion.OsloStatus.Should().Be(BuildingUnitStatus.NotRealized.Map());
+                    buildingUnitVersion.VersionTimestamp.Should().Be(buildingUnitWasNotRealized.Provenance.Timestamp);
                 });
         }
 
@@ -1454,42 +1526,47 @@ namespace BuildingRegistry.Tests.ProjectionTests.Integration
                     new Envelope<BuildingUnitWasTransferred>(new Envelope(buildingUnitWasTransferred, buildingUnitWasTransferredMetadata)))
                 .Then(async context =>
                 {
-                    var buildingUnitLatestItem =
-                        await context.BuildingUnitLatestItems.FindAsync(buildingUnitWasTransferred.BuildingUnitPersistentLocalId);
-                    buildingUnitLatestItem.Should().NotBeNull();
+                    var buildingUnitVersion =
+                        await context.BuildingUnitVersions.FindAsync(position + 1, buildingUnitWasTransferred.BuildingUnitPersistentLocalId);
+                    buildingUnitVersion.Should().NotBeNull();
 
-                    buildingUnitLatestItem!.BuildingPersistentLocalId.Should().Be(buildingUnitWasTransferred.BuildingPersistentLocalId);
-                    buildingUnitLatestItem.OsloStatus.Should().Be(
-                        BuildingUnitStatus.Parse(buildingUnitWasTransferred.Status).Map());
-                    buildingUnitLatestItem.Status.Should().Be(buildingUnitWasTransferred.Status);
-                    buildingUnitLatestItem.OsloFunction.Should().Be(
-                        BuildingUnitFunction.Parse(buildingUnitWasTransferred.Function).Map());
-                    buildingUnitLatestItem.Function.Should().Be(buildingUnitWasTransferred.Function);
-                    buildingUnitLatestItem.OsloGeometryMethod.Should().Be(BuildingUnitPositionGeometryMethod.Parse(buildingUnitWasTransferred.GeometryMethod).Map());
-                    buildingUnitLatestItem.GeometryMethod.Should().Be(buildingUnitWasTransferred.GeometryMethod);
-                    buildingUnitLatestItem.Geometry.Should().BeEquivalentTo(
+                    buildingUnitVersion!.BuildingPersistentLocalId.Should().Be(buildingUnitWasTransferred.BuildingPersistentLocalId);
+                    buildingUnitVersion.Status.Should().Be(BuildingUnitStatus.Parse(buildingUnitWasTransferred.Status).Status);
+                    buildingUnitVersion.OsloStatus.Should().Be(BuildingUnitStatus.Parse(buildingUnitWasTransferred.Status).Map());
+                    buildingUnitVersion.Function.Should().Be(BuildingUnitFunction.Parse(buildingUnitWasTransferred.Function).Function);
+                    buildingUnitVersion.OsloFunction.Should().Be(BuildingUnitFunction.Parse(buildingUnitWasTransferred.Function).Map());
+                    buildingUnitVersion.GeometryMethod.Should().Be(BuildingUnitPositionGeometryMethod.Parse(buildingUnitWasTransferred.GeometryMethod).GeometryMethod);
+                    buildingUnitVersion.OsloGeometryMethod.Should().Be(BuildingUnitPositionGeometryMethod.Parse(buildingUnitWasTransferred.GeometryMethod).Map());
+                    buildingUnitVersion.Geometry.Should().BeEquivalentTo(
                         _wkbReader.Read(buildingUnitWasTransferred.ExtendedWkbGeometry.ToByteArray()));
-                    buildingUnitLatestItem.HasDeviation.Should().BeFalse();
-                    buildingUnitLatestItem.VersionTimestamp.Should().Be(buildingUnitWasTransferred.Provenance.Timestamp);
+                    buildingUnitVersion.HasDeviation.Should().BeFalse();
+                    buildingUnitVersion.VersionTimestamp.Should().Be(buildingUnitWasTransferred.Provenance.Timestamp);
 
-                    var buildingUnitAddresses = context.BuildingUnitAddresses
-                        .Where(x => x.BuildingUnitPersistentLocalId == buildingUnitLatestItem.BuildingUnitPersistentLocalId)
-                        .ToList();
-
-                    buildingUnitAddresses.Should().HaveCount(buildingUnitWasTransferred.AddressPersistentLocalIds.Count);
+                    buildingUnitVersion.Addresses.Should().HaveCount(buildingUnitWasTransferred.AddressPersistentLocalIds.Count);
                     foreach (var addressPersistentLocalId in buildingUnitWasTransferred.AddressPersistentLocalIds)
                     {
-                        buildingUnitAddresses.SingleOrDefault(x => x.AddressPersistentLocalId == addressPersistentLocalId)
+                        buildingUnitVersion.Addresses.SingleOrDefault(x => x.AddressPersistentLocalId == addressPersistentLocalId)
                             .Should().NotBeNull();
                     }
                 });
         }
 
-        protected override BuildingUnitLatestItemProjections CreateProjection() =>
-            new BuildingUnitLatestItemProjections(new OptionsWrapper<IntegrationOptions>(new IntegrationOptions
-            {
-                BuildingNamespace = BuildingNamespace,
-                BuildingUnitNamespace = BuildingUnitNamespace
-            }));
+        protected override BuildingUnitVersionProjections CreateProjection() =>
+            new BuildingUnitVersionProjections(
+                new OptionsWrapper<IntegrationOptions>(new IntegrationOptions
+                {
+                    BuildingNamespace = BuildingNamespace,
+                    BuildingUnitNamespace = BuildingUnitNamespace,
+                }),
+                _persistentLocalIdFinder.Object,
+                new FakeAddresses());
+    }
+
+    public class FakeAddresses : IAddresses
+    {
+        public Task<int?> GetAddressPersistentLocalId(Guid addressId)
+        {
+            return Task.FromResult(1 as int?);
+        }
     }
 }
